PluginVsInstall vtk lib name and FindVTk fixes from 3.4RC (#20122)
* Fix separation character between vtk lib names and version number. (#20121)

For Mac it is '.', otherwise '-'.

* Fix config.cmake name used when searching for VTK 9 using VISIT_VTK_DIR. (#20111)

I discovered problems during a couple of my builds, where Visit was using the wrong path to VTK (not the one specified by VISIT_VTK_DIR). Finally narrowed it down to using the wrong name for VTK's installed 'config.cmake' file that helps us set VTK_DIR to point to the path containing that .cmake file.